Recalculate flow thread intrinsic size when parent needs it.
If the multicol container gets marked for intrinsic size recalculation,
its flow thread child might need to recalculate its intrinsic size as
well (the intrinsic inner size of a multicol container is stored in the
flow thread), although it may not have been marked as such. This
situation occurs when column-count changes.

Re-use the NeedsPreferredWidthsRecalculation() mechanism to fix this,
instead of inventing something new.

This fixes most of the demo in bug 1037790, except when attempting to
reduce column-count to 1.

Bug: 1037790
Change-Id: I9d42448f8cc7a5b170c98a59bcde7a916cea22db