Apply CSP to 'DOMParser.parseFromString' documents.
This applies the CSP of the context document to any doc created by calling
DOMParser.parseFromString. The original implementation, as it returned a
pointer to the CSP instance of the context document, which (in some
circumstances) could mean that the CSP of the context document itself was
modified. This makes a copy of the context document's CSP, before assigning it
to the newly created document.

This re-reverts / re-lands
  1975710: Revert CSP application to `DOMParser.parseFromString`.
  1917532: [Trusted Types] Fix Trusted Types for other document types

Bug: 1030830, 951536
Change-Id: I33adff8c376a6f37788c9aee0ef7ca0c3441785f